Update PHPVCR to avoid storing multiple fixtures
PHPVCR has updated the way SOAP headers are handled and this affects
fixture storage. We do not want to store multiple version of requests
because of this so bump the minimum version to 1.2.6 where the issue
is fixed.

This also means we can remove the now obsolete requests from our 
fixtures.

https://github.com/php-vcr/php-vcr/pull/141